Fix type issue with `SourceResource#language`
This is related to:
https://github.com/ActiveTriples/ActiveTriples/pull/159.

Adding an explicit type for Language resources deals with the problem,
in this context.

Add specs; inc. failing spec for `SourceResource#language`. The
additional specs are not meant to be exhaustive, but rather
complementary to demonstrate this is not an issue with
`DPLA::MAP::Concept` or its subclasses.

This should begin a broader effort to test this gem more effectively.