Remove LinkedVocabs
LinkedVocabs introduced a bug surrounding blank node identity, causing
new nodes to be minted in cases where the node is not already in the
cache (e.g. when the graph has been loaded from turtle).

Rather than waiting for a LinkedVocabs patch, this simply removes it,
since we aren't using the validation features anyway.